Assignment 2 – Obstacles is Mostly done
Create a Unity Tool to generate obstacles on the grid generated in Assignment 1. The Unity Tool
should have 10x10 toggleable buttons representing the grid. If the button is toggled on then that
means that particular tile is blocked using an obstacle.
The editor tool should edit a Scriptable Object in the project. Basically, the obstacle data is
stored in that scriptable object.
An ‘Obstacle Manager’ script should read the Scriptable Object and generate red spheres
representing obstacles on top of the grid created in Assignment 1.

I got it working but only during runtime.