Assignment 4 – Enemy AI Partially Done
Generate an enemy unit. The enemy unit’s objective is to move closer to the player unit. It should move using the same algorithm used in Assignment 3 or with Unity Pathfinding if Assignment 3 was not attempted. The enemy should attempt to move to one of the 4 adjacent tiles next to the player’s tile. Once it reaches the desired tile, the unit should stay still until the player unit moves. This should be done following proper OOP concepts. The ‘Enemy AI’ script is expected to inherit from an ‘AI’ interface.

Have to get the enemy to go to the correct tile